build: fix rawcopy for multiple bitbake instances
rawcopy is trying to copy binary files on a temp file with a suffix
based on partition number and create the WIC image.
If we run multiple OE instances they end up randomly creating/removing
the file causing build failures. Temporarily use a rand int as the
destrination until we fix this properly upstream

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>