make README.md small and easy to read
It's better to maintain all documentation in one place. We use
ledge-doc.git for that. Also no need to specify how to install
'repo' tool due to it's enabled in linux distributions now. No
need to discuss about fTPM hack due to we upsteamed it. So
just keep this file minimal with the reference to more detailed
documentation.

Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>